+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ﻿## Enhancements
-* Bump sweetalert2 dependency to 8.15.3
-  * Bug fix: expand/shrink popup accordingly to textarea width (https://github.com/sweetalert2/sweetalert2/pull/1702) (https://github.com/sweetalert2/sweetalert2/commit/93f59dc)
+* Bump @sweetalert2/themes dependency to 2.1.0
+  * 🎉 New Theme: Bootstrap 4 🎉 (https://github.com/sweetalert2/sweetalert2-themes/issues/23) (https://github.com/sweetalert2/sweetalert2-themes/commit/a802ed3)
+* Removed lodash flatten function to flatten queue results. Reduced bundle size.